feat(rostering): make roster grid the rostering home (Stage 1 routing)
Grid-first IA: /rostering/[group_id] now renders RosterGridScreen instead of
the Schedule list. The (hub) layout drops its shared header + JS top tab bar
(now a bare Slot); the grid supplies its own header. Standalone Teams /
Cross-team screens gain their own back header (RosteringBackHeader). /grid
redirects to the home so there is one grid URL.
