+++ apps/mobile/app/rostering/[group_id]/(hub)/_layout.tsx
@@ -1,73 +1,15 @@
-import React from "react";
-import { View, Text, StyleSheet, TouchableOpacity } from "react-native";
-import { Slot, useRouter } from "expo-router";
-import { useSafeAreaInsets } from "react-native-safe-area-context";
-import { Ionicons } from "@expo/vector-icons";
-import { useTheme } from "@hooks/useTheme";
-import { RosteringTopTabBar } from "@features/scheduling";
+import { Slot } from "expo-router";
 
 /**
- * Rostering hub chrome.
+ * Rostering home layout.
  *
- * Renders the shared header and the JS-only top tab bar above whichever hub
- * tab is active (Schedule / Teams / Cross-team, rendered into `<Slot />`).
- * Event and team detail screens are NOT under this layout — they push over
- * the whole hub. See ADR-024.
+ * Grid-first IA (Stage 1): the roster grid is the rostering home and renders
+ * its own header, so this layout is now a bare pass-through. The old hub
+ * chrome (shared "Rostering" header + JS top tab bar) is gone — Teams and
+ * Cross-team are reached from the grid's ⋯ overflow and render their own
+ * back headers. The `(hub)` route group itself is removed in a later cleanup
+ * stage. See ADR-024.
  */
-export default function RosteringHubLayout() {
-  const { colors } = useTheme();
-  const insets = useSafeAreaInsets();
-  const router = useRouter();
-
-  return (
-    <View
-      style={[
-        styles.container,
-        { paddingTop: insets.top, backgroundColor: colors.surface },
-      ]}
-    >
-      <View style={[styles.header, { borderBottomColor: colors.border }]}>
-        <TouchableOpacity
-          onPress={() => {
-            if (router.canGoBack()) router.back();
-          }}
-          hitSlop={12}
-          style={styles.headerSide}
-        >
-          <Ionicons name="chevron-back" size={28} color={colors.text} />
-        </TouchableOpacity>
-        <Text style={[styles.headerTitle, { color: colors.text }]}>
-          Rostering
-        </Text>
-        <View style={styles.headerSide} />
-      </View>
-
-      <RosteringTopTabBar />
-
-      <Slot />
-    </View>
-  );
+export default function RosteringHomeLayout() {
+  return <Slot />;
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  header: {
-    flexDirection: "row",
-    alignItems: "center",
-    paddingHorizontal: 8,
-    paddingVertical: 12,
-    borderBottomWidth: StyleSheet.hairlineWidth,
-  },
-  headerSide: {
-    width: 36,
-    padding: 4,
-  },
-  headerTitle: {
-    flex: 1,
-    fontSize: 17,
-    fontWeight: "600",
-    textAlign: "center",
-  },
-});
