+++ apps/mobile/app/rostering/[group_id]/(hub)/index.tsx
@@ -1,6 +1,12 @@
-import { EventListScreen } from "@features/scheduling";
+import { RosterGridScreen } from "@features/scheduling";
 
-/** Schedule tab of the Rostering hub — route `/rostering/[group_id]`. */
-export default function RosteringScheduleTab() {
-  return <EventListScreen />;
+/**
+ * Rostering home — route `/rostering/[group_id]`.
+ *
+ * Grid-first IA (Stage 1): the roster grid is the rostering home. It renders
+ * its own header and hosts Teams / Cross-team / Collect availability in a ⋯
+ * overflow. The old Schedule list (EventListScreen) is no longer the home.
+ */
+export default function RosteringHome() {
+  return <RosterGridScreen />;
 }
