+++ apps/mobile/app/rostering/[group_id]/_layout.tsx
@@ -3,9 +3,9 @@ import { Stack } from "expo-router";
 /**
  * Group-scoped rostering navigator.
  *
- * The `(hub)` route group is the tabbed Rostering hub (Schedule / Teams /
- * Cross-team); `event` and `team` detail screens push over it as plain
- * stack screens. See ADR-024.
+ * Grid-first IA (Stage 1): the rostering home (`(hub)/index`) renders the
+ * roster grid. Teams / Cross-team and the `event` / `team` / `availability`
+ * detail screens push over it as plain stack screens. See ADR-024.
  */
 export default function RosteringGroupLayout() {
   return <Stack screenOptions={{ headerShown: false }} />;
